+++ ninaAPI/WebService/V2/Equipment/FlatDevice.cs
@@ -91,58 +91,6 @@ namespace ninaAPI.WebService.V2
             HttpContext.WriteToResponse(response);
         }
 
-        [Route(HttpVerbs.Get, "/equipment/flatdevice/connect")]
-        public async Task FlatDeviceConnect([QueryField] bool skipRescan)
-        {
-            HttpResponse response = new HttpResponse();
-
-            try
-            {
-                IFlatDeviceMediator flat = AdvancedAPI.Controls.FlatDevice;
-
-                if (!flat.GetInfo().Connected)
-                {
-                    if (!skipRescan)
-                    {
-                        await flat.Rescan();
-                    }
-                    await flat.Connect();
-                }
-                response.Response = "Flatdevice connected";
-            }
-            catch (Exception ex)
-            {
-                Logger.Error(ex);
-                response = CoreUtility.CreateErrorTable(CommonErrors.UNKNOWN_ERROR);
-            }
-
-            HttpContext.WriteToResponse(response);
-        }
-
-        [Route(HttpVerbs.Get, "/equipment/flatdevice/disconnect")]
-        public async Task FlatDeviceDisconnect()
-        {
-            HttpResponse response = new HttpResponse();
-
-            try
-            {
-                IFlatDeviceMediator flat = AdvancedAPI.Controls.FlatDevice;
-
-                if (flat.GetInfo().Connected)
-                {
-                    await flat.Disconnect();
-                }
-                response.Response = "Flatdevice disconnected";
-            }
-            catch (Exception ex)
-            {
-                Logger.Error(ex);
-                response = CoreUtility.CreateErrorTable(CommonErrors.UNKNOWN_ERROR);
-            }
-
-            HttpContext.WriteToResponse(response);
-        }
-
         [Route(HttpVerbs.Get, "/equipment/flatdevice/set-light")]
         public void FlatDeviceToggle([QueryField] bool on)
         {